pyproject.toml
[tool.poetry]
name = "humpback-whale"
version = "0.1.0"
description = "A package to train and infer the solution on \"Humpback Whale Identification\" competition on Kaggle."
authors = ["Denis Ivanov <denivsemik@gmail.com>"]
readme = "README.md"
packages = [
    { include = "humpback_whale" }
]

[tool.poetry.scripts]
whale = "humpback_whale.__main__:main"

[tool.poetry.dependencies]
python = "^3.13"
lightning = "^2.5.1.post0"
timm = "^1.0.15"
mlflow = "^3.1.0"
torchmetrics = "^1.7.3"
pandas = "^2.3.0"
torchvision = "^0.22.1"
hydra-core = "^1.3.2"
fire = "^0.7.0"
numpy = "^2.3.0"
dvc = "^3.60.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
